refactor: split cli/_helpers.py into _helpers/ package
