+++ initrunner/cli/_helpers/_console.py
@@ -1,0 +1,5 @@
+"""Shared Rich console singleton."""
+
+from rich.console import Console
+
+console = Console()
